date fixes + clean up

+++ src/lib/utils.h
@@ -28,7 +28,7 @@ extern "C" {
 // Converts data bytes into a hexstring !! IN PLACE !!
 uint8_t convertToHexstringInPlace(uint8_t *data, uint16_t dataLen, uint16_t dataLenMax);
 
-void printTime(char *out, uint16_t outLen, uint32_t t);
+void printTime(char *out, uint16_t outLen, uint64_t t);
 
 #ifdef __cplusplus
 }
